tools/scripts/yaml2sheet/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yaml2sheet"
version = "0.2.0"
description = "Generate checklist in Google Sheets from YAML data"
readme = "README.md"
authors = [
    { name = "Masafumi NAKANE", email = "max@freee.co.jp" }
]
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "freee_a11y_gl>=0.2.2"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    # Project-specific dev dependencies only
    # Common dev tools (pytest, black, mypy, etc.) are managed at root level
    # Install with: pip install -r requirements-dev.txt
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["yaml2sheet*"]

[project.scripts]
yaml2sheet = "yaml2sheet.yaml2sheet:main"
